Pure rename: change `Selfie.FooSelfie` to to be just `FooSelfie`.
