+++ jvm/selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/coroutines/Coroutines.kt
@@ -16,9 +16,9 @@
 package com.diffplug.selfie.coroutines
 
 import com.diffplug.selfie.Camera
+import com.diffplug.selfie.DiskSelfie
 import com.diffplug.selfie.Roundtrip
 import com.diffplug.selfie.RoundtripJson
-import com.diffplug.selfie.Selfie
 import com.diffplug.selfie.Snapshot
 import com.diffplug.selfie.guts.CoroutineDiskStorage
 import kotlin.coroutines.coroutineContext
@@ -46,7 +46,7 @@ private suspend fun disk() =
 suspend fun <T> expectSelfie(actual: T, camera: Camera<T>) = expectSelfie(camera.snapshot(actual))
 suspend fun expectSelfie(actual: String) = expectSelfie(Snapshot.of(actual))
 suspend fun expectSelfie(actual: ByteArray) = expectSelfie(Snapshot.of(actual))
-suspend fun expectSelfie(actual: Snapshot) = Selfie.DiskSelfie(actual, disk())
+suspend fun expectSelfie(actual: Snapshot) = DiskSelfie(actual, disk())
 suspend fun preserveSelfiesOnDisk(vararg subsToKeep: String) {
   val disk = disk()
   if (subsToKeep.isEmpty()) {
